user profile update, user-connection updates
